pyproject.toml
[tool.poetry]
name = "weather-data-etl-pipeline"
version = "0.1.0"
description = ""
authors = ["sch-paulo <pschmitzjr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openmeteo-requests = "^1.3.0"
requests-cache = "^1.2.1"
retry-requests = "^2.0.0"
numpy = "^2.2.1"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
